Reversion to v2_beta1; reorganisation of notebook
Decided change in soil & soluble P equations is added complexity for no
additional benefit, therefore reverted to v2_beta1.
User-input that was previously in the ipython notebook now moved to the
parameter file ('Setup' worksheet), apart from the location of the
parameter file.
Chunks of code moved to within the model function
Started generalising the writing of output & graphs (not finished).