beta 4: New simpler sediment equations
After model testing, decided to simplify the model by removing the
M_land term (dependent on quick flow), so that the only sediment inputs
to the reach are dependent on in-stream flow. Re-formulated the old
entrainment term so that the scaling factor incorporates all the
terrestrial erodibility factors.

Also added in markdown cells to improve model readability.